refact: big refact
