+++ Makefile
@@ -67,8 +67,5 @@ migrate-down:
 lint:
 	golangci-lint run ./...
 
-swag:
-	swag init -g main.go --output docs
-
 check:
 	staticcheck ./...
