+++ README.md
@@ -37,17 +37,12 @@ $ make dev
 # Run the migrations
 $ make migrate-up
 
-# Generate documentation
-$ make swag
-
 # Run the application in development mode
 $ go run main.go
 ```
 
 ```bash
 - The backend will open on the port:5000        # access http://localhost:5000
-- The mailcatcher will open on the port:1080    # access http://localhost:1080
-- The documentation will open on the port:5000: # access http://localhost:5000/swagger/index.html
 ```
 ## 😯 How to contribute to the project
