+++ docker-compose.dev.yml
@@ -16,16 +16,6 @@ services:
     networks:
       - powersso
     
-  mailcatcher:
-    container_name: mailcatcher
-    image: sj26/mailcatcher
-    expose:
-      - 1025
-      - 1080
-    ports:
-      - 1025:1025
-      - 1080:1080
-
 networks:
   powersso:
     driver: bridge
