+++ domain/auth/service.go
@@ -4,6 +4,7 @@ import (
 	"encoding/base32"
 
 	"github.com/google/uuid"
+	"github.com/isaqueveras/powersso/config"
 	"github.com/isaqueveras/powersso/utils"
 )
 
@@ -31,7 +32,22 @@ func (s *Service) Configure2FA(userID *uuid.UUID) (err error) {
 	data := []byte(utils.RandomString(26))
 	dst := make([]byte, base32.StdEncoding.EncodedLen(len(data)))
 	base32.StdEncoding.Encode(dst, data)
-	return s.repoOTP.SetToken(utils.Pointer(string(dst)))
+	return s.repoOTP.SetToken(userID, utils.Pointer(string(dst)))
 }
 
-func (*Service) GenerateQrCode2FA() {}
+// GenerateQrCode2FA return the formatted url to configure 2-factor authentication
+func (s *Service) GenerateQrCode2FA(userID *uuid.UUID) (url *string, err error) {
+	userName, token, err := s.repoOTP.GetToken(userID)
+	if err != nil {
+		return nil, err
+	}
+
+	if config.Get().Server.IsModeDevelopment() {
+		*userName += " [DEV]"
+	}
+
+	projectName := utils.Pointer(config.Get().Meta.ProjectName)
+	link := utils.GetUrlQrCode(projectName, token, userName)
+
+	return utils.Pointer(link), nil
+}
