+++ infrastructure/persistencie/auth/repository.go
@@ -1,0 +1,118 @@
+// Copyright (c) 2023 Isaque Veras
+// Use of this source code is governed by a MIT style
+// license that can be found in the LICENSE file.
+
+package auth
+
+import (
+	"github.com/google/uuid"
+	database "github.com/isaqueveras/powersso/database/postgres"
+	domain "github.com/isaqueveras/powersso/domain/auth"
+	infra "github.com/isaqueveras/powersso/infrastructure/persistencie/auth/postgres"
+)
+
+var (
+	_ domain.IOTP     = (*repoOTP)(nil)
+	_ domain.IFlag    = (*repoFlag)(nil)
+	_ domain.IAuth    = (*repoAuth)(nil)
+	_ domain.ISession = (*repoSession)(nil)
+	_ domain.IUser    = (*repoUser)(nil)
+)
+
+type (
+	repoOTP     struct{ pg *infra.OTP }
+	repoFlag    struct{ pg *infra.Flag }
+	repoAuth    struct{ pg *infra.PGAuth }
+	repoSession struct{ pg *infra.Session }
+	repoUser    struct{ pg *infra.User }
+)
+
+// NewAuthRepository creates a new repository
+func NewAuthRepository(tx *database.Transaction) domain.IAuth {
+	return &repoAuth{pg: &infra.PGAuth{DB: tx}}
+}
+
+// CreateAccount contains the flow for the user register in database
+func (r *repoAuth) CreateAccount(data *domain.CreateAccount) (userID *uuid.UUID, err error) {
+	return r.pg.CreateAccount(data)
+}
+
+// AddAttempts contains the flow for the add number failed attempts
+func (r *repoAuth) AddAttempts(userID *uuid.UUID) error {
+	return r.pg.AddAttempts(userID)
+}
+
+// LoginSteps contains the flow to get the data needed to retrieve the steps required to log in a user
+func (r *repoAuth) LoginSteps(email *string) (*domain.Steps, error) {
+	return r.pg.LoginSteps(email)
+}
+
+// NewOTPRepo creates a new repository
+func NewOTPRepo(tx *database.Transaction) domain.IOTP {
+	return &repoOTP{pg: &infra.OTP{DB: tx}}
+}
+
+func (r *repoOTP) GetToken(userID *uuid.UUID) (*string, *string, error) {
+	return r.pg.GetToken(userID)
+}
+
+func (r *repoOTP) SetToken(userID *uuid.UUID, secret *string) error {
+	return r.pg.SetToken(userID, secret)
+}
+
+// NewFlagRepo creates a new repository
+func NewFlagRepo(tx *database.Transaction) domain.IFlag {
+	return &repoFlag{pg: &infra.Flag{DB: tx}}
+}
+
+func (r *repoFlag) Set(userID *uuid.UUID, flag domain.Flag) error {
+	return r.pg.Set(userID, flag)
+}
+
+func (r *repoFlag) Get(userID *uuid.UUID) (*int64, error) {
+	return r.pg.Get(userID)
+}
+
+// NewSessionRepository creates a new repository
+func NewSessionRepository(tx *database.Transaction) domain.ISession {
+	return &repoSession{pg: &infra.Session{DB: tx}}
+}
+
+// Create create a new session for a user
+func (r *repoSession) Create(userID *uuid.UUID, clientIP, userAgent *string) (*uuid.UUID, error) {
+	return r.pg.Create(userID, clientIP, userAgent)
+}
+
+// Delete delete a session for a user
+func (r *repoSession) Delete(ids ...*uuid.UUID) error {
+	return r.pg.Delete(ids...)
+}
+
+func (r *repoSession) Get(userID *uuid.UUID) ([]*uuid.UUID, error) {
+	return r.pg.Get(userID)
+}
+
+// NewUserRepository creates a new repository
+func NewUserRepository(tx *database.Transaction) domain.IUser {
+	return &repoUser{pg: &infra.User{DB: tx}}
+}
+
+// Get get user data
+func (r *repoUser) Get(user *domain.User) error {
+	return r.pg.Get(user)
+}
+
+// Exist check if user already exists
+func (r *repoUser) Exist(email *string) error {
+	return r.pg.Exist(email)
+}
+
+// Disable deactivate a user's account
+func (r *repoUser) Disable(userUUID *uuid.UUID) error {
+	return r.pg.Disable(userUUID)
+}
+
+// ChangePassword manages the flow to change a user's password
+func (r *repoUser) ChangePassword(in *domain.ChangePassword) error {
+	return r.pg.ChangePassword(in)
+}
