--- infrastructure/persistencie/auth/user_repository.go
@@ -1,40 +1,0 @@
-// Copyright (c) 2023 Isaque Veras
-// Use of this source code is governed by a MIT style
-// license that can be found in the LICENSE file.
-
-package auth
-
-import (
-	"github.com/google/uuid"
-	pg "github.com/isaqueveras/powersso/database/postgres"
-	domain "github.com/isaqueveras/powersso/domain/auth"
-	infra "github.com/isaqueveras/powersso/infrastructure/persistencie/auth/postgres"
-)
-
-var _ domain.IUser = (*repoUser)(nil)
-
-type repoUser struct{ pg *infra.PGUser }
-
-// NewUserRepository creates a new repository
-func NewUserRepository(tx *pg.Transaction) domain.IUser {
-	return &repoUser{pg: &infra.PGUser{DB: tx}}
-}
-
-// Get get user data
-func (r *repoUser) Get(user *domain.User) error {
-	return r.pg.Get(user)
-}
-
-// Exist check if user already exists
-func (r *repoUser) Exist(email *string) error {
-	return r.pg.Exist(email)
-}
-
-// Disable deactivate a user's account
-func (r *repoUser) Disable(userUUID *uuid.UUID) error {
-	return r.pg.Disable(userUUID)
-}
-
-func (r *repoUser) ChangePassword(in *domain.ChangePassword) error {
-	return r.pg.ChangePassword(in)
-}
